refactor(core): Bus / Peripheral / LinearMemory address types u64 → u32
ARM Cortex-M and RV32 both have 32-bit address spaces. The Bus and
Peripheral traits, LinearMemory, DmaRequest, and SimulationError all
carried `u64` addresses, forcing every call site to `... as u64` and
cluttering signatures. The roast flagged this explicitly.

Converted:

  SimulationError::MemoryViolation(u64)  →  (u32)
  SimulationError::DecodeError(u64)      →  (u32)
  DmaRequest::addr: u64                  →  u32
  PeripheralEntry::{base, size}: u64     →  u32
  Bus::{read,write}_u8/u16/u32(addr: u64) → u32
  Peripheral::{read,write}(offset: u64)  →  u32
  LinearMemory::{new, read_u8, write_u8, base_addr, Segment::start_addr,
    ProgramImage::{new, entry_point, add_segment}} — all u64 → u32.
  RiscV + CortexM bus calls + frame-pointer arithmetic: dropped the
    `... as u64` casts wholesale.
  GenericPeripheral: register-offset comparison upcast from `u32`
    from u64 (descriptor stays u64 because YAML numbers are u64).

Callers that previously wrote `(X) as u64` to cross the Bus boundary
now pass `u32` values directly. 118 such casts removed.

At the YAML → bus boundary (crates/core/src/bus/mod.rs::from_config)
the config crate still exposes `u64` addresses (since YAML numbers
deserialise as u64). We narrow at that boundary with a single
`as u32` cast per field — u32 is sufficient for any 32-bit MCU and
flagging overflow there is cheaper than polluting the whole trait
surface.

Tests (per-iteration real-firmware policy — this is a wide blast
radius refactor, so FW suite is the real proof):
  cargo test -p labwired-core --lib       — 95 passed, 0 failed.
  cargo test -p labwired-cli --tests      — 35 passed, 0 failed
                                            (real ELF ran through
                                            end-to-end — every
                                            bus read/write is u32 now).
  cargo check --workspace                 — 0 warnings.

+++ crates/cli/src/main.rs
@@ -208,8 +208,8 @@ use labwired_core::snapshot::CpuSnapshot;
 #[derive(Debug, Serialize, Deserialize)]
 struct PeripheralSnapshot {
     name: String,
-    base: u64,
-    size: u64,
+    base: u32,
+    size: u32,
     irq: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     state: Option<serde_json::Value>,
