+++ crates/core/src/peripherals/dma.rs
@@ -29,7 +29,7 @@ impl Dma1 {
         Self::default()
     }
 
-    fn read_reg(&self, offset: u64) -> u32 {
+    fn read_reg(&self, offset: u32) -> u32 {
         match offset {
             0x00 => self.isr,
             _ => {
@@ -50,7 +50,7 @@ impl Dma1 {
         }
     }
 
-    fn write_reg(&mut self, offset: u64, value: u32) {
+    fn write_reg(&mut self, offset: u32, value: u32) {
         match offset {
             0x04 => {
                 // IFCR: Write 1 to clear corresponding ISR bits
@@ -81,14 +81,14 @@ impl Dma1 {
 }
 
 impl Peripheral for Dma1 {
-    fn read(&self, offset: u64) -> SimResult<u8> {
+    fn read(&self, offset: u32) -> SimResult<u8> {
         let reg_offset = offset & !3;
         let byte_offset = (offset % 4) as u32;
         let reg_val = self.read_reg(reg_offset);
         Ok(((reg_val >> (byte_offset * 8)) & 0xFF) as u8)
     }
 
-    fn write(&mut self, offset: u64, value: u8) -> SimResult<()> {
+    fn write(&mut self, offset: u32, value: u8) -> SimResult<()> {
         let reg_offset = offset & !3;
         let byte_offset = (offset % 4) as u32;
 
@@ -121,7 +121,7 @@ impl Peripheral for Dma1 {
                 if dir_bit == 1 {
                     // Memory to Peripheral
                     dma_requests.push(DmaRequest {
-                        addr: chan.cmar as u64,
+                        addr: chan.cmar,
                         val: 0, // Value will be populated by SystemBus during Read
                         direction: DmaDirection::Read,
                     });
@@ -145,7 +145,7 @@ impl Peripheral for Dma1 {
 
                         // Let's just implement a fake write for now to verify the plumbing.
                         dma_requests.push(DmaRequest {
-                            addr: chan.cmar as u64,
+                            addr: chan.cmar,
                             val: 0x42, // Dummy value
                             direction: DmaDirection::Write,
                         });
