+++ crates/core/src/peripherals/exti.rs
@@ -29,7 +29,7 @@ impl Exti {
         }
     }
 
-    fn read_reg(&self, offset: u64) -> u32 {
+    fn read_reg(&self, offset: u32) -> u32 {
         match offset {
             0x00 => self.imr,
             0x04 => self.emr,
@@ -41,7 +41,7 @@ impl Exti {
         }
     }
 
-    fn write_reg(&mut self, offset: u64, value: u32) {
+    fn write_reg(&mut self, offset: u32, value: u32) {
         match offset {
             0x00 => self.imr = value & 0x7FFFF,
             0x04 => self.emr = value & 0x7FFFF,
@@ -64,14 +64,14 @@ impl Exti {
 }
 
 impl Peripheral for Exti {
-    fn read(&self, offset: u64) -> SimResult<u8> {
+    fn read(&self, offset: u32) -> SimResult<u8> {
         let reg_offset = offset & !3;
         let byte_offset = (offset % 4) as u32;
         let reg_val = self.read_reg(reg_offset);
         Ok(((reg_val >> (byte_offset * 8)) & 0xFF) as u8)
     }
 
-    fn write(&mut self, offset: u64, value: u8) -> SimResult<()> {
+    fn write(&mut self, offset: u32, value: u8) -> SimResult<()> {
         let reg_offset = offset & !3;
         let byte_offset = (offset % 4) as u32;
